backend url added for login and register. added  favicon and updated title for app
