+++ src/components/Register.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import { Redirect, Link } from 'react-router-dom';
 import logo from '../logo-complete.svg'
+import Api from '../services/api';
 
 class Register extends React.Component {
   constructor(props){
@@ -25,18 +26,8 @@ class Register extends React.Component {
       password: this.state.password
     }}
 
-    const reqObj = {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json',
-        Accept: 'application/json'
-      },
-      body: JSON.stringify(bodyObj)
-    }
 
-    fetch('http://localhost:3000/api/v1/users', reqObj)
-      .then(resp => resp.json())
-      .then(user => {
+      Api.createUser(bodyObj).then(user => {
         if(!user['error']){
           this.props.getUser(user)
         }else {
